[Step4] [fix] Refactor
