+++ src/main/java/racingCarGame/util/RandomGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package racingCarGame.util;
 
 import java.util.Random;
