--- src/test/java/domain/AlwaysMoveStrategy.java
@@ -1,8 +1,0 @@
-package domain;
-
-public class AlwaysMoveStrategy implements MoveStrategy{
-    @Override
-    public boolean isCanMove() {
-        return true;
-    }
-}
