+++ src/test/java/racingCarGame/domain/AlwaysMoveStrategy.java
@@ -1,0 +1,10 @@
+package racingCarGame.domain;
+
+import racingCarGame.domain.MoveStrategy;
+
+public class AlwaysMoveStrategy implements MoveStrategy {
+    @Override
+    public boolean isCanMove() {
+        return true;
+    }
+}
