+++ src/test/java/racingCarGame/domain/CarTest.java
@@ -1,4 +1,4 @@
-package domain;
+package racingCarGame.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
